Condensed post scoring modals into one file
